+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Changed
 - Neo input blocks now have a `data-neo-b-name` attribute (thanks @davidhellmann)
+- Block type descriptions can appear on hover over new block buttons
 
 ### Fixed
 - Fixed a bug where copied top level Neo blocks could be pasted into places in the Neo field where that block type shouldn't have been allowed